+++ packages/vexide-devices/src/adi/encoder.rs
@@ -249,11 +249,11 @@ impl AdiEncoder {
     }
 }
 
-impl AdiDevice for AdiEncoder {
+impl AdiDevice<2> for AdiEncoder {
     type PortNumberOutput = (u8, u8);
 
-    fn port_number(&self) -> Self::PortNumberOutput {
-        (self.top_port.number(), self.bottom_port.number())
+    fn port_numbers(&self) -> [u8; 2] {
+        [self.top_port.number(), self.bottom_port.number()]
     }
 
     fn expander_port_number(&self) -> Option<u8> {