+++ packages/vexide-devices/src/adi/motor.rs
@@ -227,11 +227,11 @@ impl AdiMotor {
     }
 }
 
-impl AdiDevice for AdiMotor {
+impl AdiDevice<1> for AdiMotor {
     type PortNumberOutput = u8;
 
-    fn port_number(&self) -> Self::PortNumberOutput {
-        self.port.number()
+    fn port_numbers(&self) -> [u8; 1] {
+        [self.port.number()]
     }
 
     fn expander_port_number(&self) -> Option<u8> {