+++ packages/vexide-devices/src/adi/potentiometer.rs
@@ -160,11 +160,11 @@ impl PotentiometerType {
     }
 }
 
-impl AdiDevice for AdiPotentiometer {
+impl AdiDevice<1> for AdiPotentiometer {
     type PortNumberOutput = u8;
 
-    fn port_number(&self) -> Self::PortNumberOutput {
-        self.port.number()
+    fn port_numbers(&self) -> [u8; 1] {
+        [self.port.number()]
     }
 
     fn expander_port_number(&self) -> Option<u8> {