+++ packages/vexide-devices/src/adi/pwm.rs
@@ -94,11 +94,11 @@ impl AdiPwmOut {
     }
 }
 
-impl AdiDevice for AdiPwmOut {
+impl AdiDevice<1> for AdiPwmOut {
     type PortNumberOutput = u8;
 
-    fn port_number(&self) -> Self::PortNumberOutput {
-        self.port.number()
+    fn port_numbers(&self) -> [u8; 1] {
+        [self.port.number()]
     }
 
     fn expander_port_number(&self) -> Option<u8> {