+++ packages/vexide-devices/src/adi/servo.rs
@@ -138,11 +138,11 @@ impl AdiServo {
     }
 }
 
-impl AdiDevice for AdiServo {
+impl AdiDevice<1> for AdiServo {
     type PortNumberOutput = u8;
 
-    fn port_number(&self) -> Self::PortNumberOutput {
-        self.port.number()
+    fn port_numbers(&self) -> [u8; 1] {
+        [self.port.number()]
     }
 
     fn expander_port_number(&self) -> Option<u8> {