+++ CHANGELOG.md
@@ -40,6 +40,8 @@ Before releasing:
 - All `Position` methods are now usable in `const` context. (#254)
 - Two-wire ADI devices (`AdiEncoder` and `AdiRangeFinder`) now take their ports as separate arguments instead of a tuple. (#271) (**Breaking Change**)
 - `AdiEncoder` and `AdiRangeFinder` will now panic if invalid port pairings are passed rather than return a `Result`. (#271) (**Breaking Change**)
+- `AdiDevice` is now const-generic over the number of ports used by the device. (#271) (**Breaking Change**)
+- Replaced `AdiDevice::port_number` with `AdiDevice::port_numbers`. (#271) (**Breaking Change**)
 
 ### Removed